Comprehensive reading of FC schema
Added a new method to FeatureJSON to obtain the a feature collection schema by reading all features in search of the data types of all properties.

FeatureJSON#readFeatureCollection schema reads the {@link SimpleFeatureType} of a GeoJSON using FeatureTypeHandler, which is simpler handler when compared to FeatureHandler. In the  worst case, it will parse all features searching for attributes not present in previous features.