Add support for the citext extension in PostGIS.
This change allows the citext type to be used in features, allowing indexed case
insensitive queries. There probably could be better handling for extension types
in general for the PostGISDialect, but I'm not entirely sure what that design
would be, so I just followed the pattern set by the uuid type.