Separate ShapefileDataStoreFactory and ShapefileDirectoryFactory
The two was written in a strange way. This commit separates the two
factories. One operating single shapefiles. The other operating
directories of shapefiles. The Param file-type was removed from
ShapeFileDataStoreFactory. This is given by the extension of the file
and wether it's a directory or file.
ShapefileDirectoryFactory uses ShapefileDataStoreFactory as a mean to
create ShapefileDataStore's.