Fix broken JDBCJoinTests, Fixed JDBCDataStore bug
Before this fix the tests in JDBCJoinTest didn't test the
exposePrimaryKeys flag in order to determine the number of attributes
expected. Making that change exposed the fact that JDBCDataStore didn't
invalidate cached FeatureType instances in it's FeatureStore's
ContentEntry's JDBCState objects. This was causing modifications to
JDBCDataStore.exposePrimaryKeyColumns to not change the primary key
exposition behavior of FeatureTypes which had been read through the
JDBCDataStore prior to the modification.

Deleted strange assertions from JDBCJoinTest.testOuterJoin, and
JDBCJoinTest.testJoinMoreThanTwo.

Finally, the JDBCDataStore bug was masking an issue with
JDBCJoinTest.testSpatialJoin, as this test was setting a list of
PropertyNames in its queries, however this is presently not supported
with join queries.