KML22 - dump feature collections
Following this thread on the ML:
http://sourceforge.net/p/geotools/mailman/message/33052373/

This commit aims to add write support of feature collections into
KML2.2. It also allows to dump attributes as schema-less ExtendedData.

Manual tests, as well as a new tests have been added to check that the
produced document is correct and that a generated one can be parsed.

Backort of #604.