[GEOT-5289] Use java.util.ServiceLoader instead of javax.imageio.spi.ServiceRegistry
Update code to rely on java.util.ServiceLoader instead of
javax.imageio.spi.ServiceRegistry, because of a behavioral change of
Java 9 which no longer allows to load service providers of other than
Image I/O service types. Because of this change, GeoTools does not work
with Java 9 but causes an IllegalArgumentException during
initialization.