GEOT-6018 - Fix for WMTS doesn't correctly handle `image/jpgpng` format
Centralized parameter FORMAT for GetTile Requests to find out easily
when it has been set.
Setting this parameter by the coverage reader as it has been computed
there already in constructor.
Set this value also for GetTileRequest when working with KVP.

My provided test is set to ignore, as it depends on online data. I do
not know how to test this the right way in GEOTOOLS.