[GEOT-6202] gt-jdbc-hana - Support for HANA 1
Currently, the gt-jdbc-hana plugin supports HANA 2 only.

If used with a HANA 1 instance, some queries work, but many fail
because some functions are not available in HANA 1 yet.

With this change, gt-jdbc-hana supports the revisions of HANA 1 that
are still supported by SAP (120 and later). gt-jdbc-hana will reject
older revisions (< 120) immediately on establishing a connection.