[GEOT-6358] xsd-kml: Fix LatLonAltBox parsing for KML 2.2
Add explicit binding for LatLonAltBoxType in KML 2.2
Configuration. Otherwise parsing a Region with a
LatLonAltBox child element throws a ClassCastException
because it wasn't parsed to a JTS Envelope.

Reason being that in KML 2.1 LatLonAltBoxType is a subtype
of LatLonBoxType, but in KML 2.2 it is not, both being
derived from AbstractLatLonBoxType instead.