Remove jdom dependency from main and shapefile modules
there's a direct dependency on the jdom library
from main and shapefile, which adds nothing that
can't be done with official XML API's, hence
drawing the dependency unnnecessary.

This patch removes such dependency and replaces the
XML parsing done with it by the StAX API which is
faster and is part of the default Java library.

For instance, the following classes have been adapted:

* org.geotools.data.ows.ServiceExceptionParser
* org.geotools.data.shapefile.shp.xml.ShpXmlFileReader