[GEOT-6422] Improve performance of JDBCDataStoreFactory.isAvailable()
`JDBCDataStoreFactory.isAvailable()` uses
`Class.forName(getDriverClassName())` to determine availability based on
the presence of the JDBC driver in the classpath.

`Class.forName()` is the cuase of a great deal of thread contention and
should be avoided as much as possible in highly concurrent code, which
is the case of `DataAccessFactory.isAvailable()` at least in the context
of GeoServer.

This patch computes and caches the availability of each driver *per instance*,
keeping the calls to `Class.forName` serialized in order to avoid potential
deadlocks caused by some drivers' registration to DriverManager as part
of their static initialization.