GEOT-6422: Fix FactoryRegistry.getFactories and get unsorted DataAccessFactories
- `DataAccessFactory.getAvailableDataStores` does not need to retrieve
the factories in any particular order, so better call the version of
`FactoryRegistry.getFactories` that receives a boolean flag indicating
whether the prescribed factory order is to be returned with `false` as
value, and save the performance penalty introduced by the
`TopologicalSortIterator` further down the call chain.

- These two overloaded versions of `FactoryRegistry.getFactories` are
supposed to behave the same except that `getFactories(Class, Predicate,
Hints)` composes an "isAvailable" predicate (which in the case of
`DataAccessFactory` and `DataStoreFactorySpi` is irrelevant) to then
delegate to `getFactories(Class, Predicate, boolean useOrdering)`.

Problem is only the former is initializing the available "categories" by
means of `synchronizeIteratorProviders()` and
`scanForPluginsIfNeeded(category)`, while both should, hence this fixes
it by moving the initialization call to the actual method tha retrieves
the factories from the registry.