Remove OraclePrimaryKeyFinderTestSetup#testSequencedPrimaryKey, it's breaking the build on Travis.
Before the removal we have tried to reproduce the issue on two machines,
failing to do so, and tried to analyze it running builds on a branch,
but it did not occurr there either (only difference, a loop printing
features before running the test). For some reason the sequence backing
seqtable gets moved up to 21, but we found no way to make it happen.
The test involved is also the only one using seqtable and pksequence.
A true mistery.