[GEOT-6800] Certain shapefiles can still trigger exceptions when Z value present but no M value
GEOT-6740 fixed a buffer underflow exception when reading shapefile
data - the problem was the code tried to read the M-value data
regardless of whether M data was present in the shapefile.

However, the fix was incomplete - it only checked there was one more
double value remaining. The code would actually try to read a lot more M
data than that. An example would be if the buffer contained a large
shape (e.g. a circle) that was followed by a much smaller shape (e.g. a
rectangle). In this case there's still more data remaining in the buffer,
but not enough to match the numPoints in the shape.

The fix is to check there as *at least* the expected amount of M-value
data still remaining in the buffer before trying to read it.

Test data files were updated so that they trigger the exception. Note
that the order of shapes can be important - we usually want a large
shape followed by a small shape.
- zpolygons.shp: added a circle followed by a rectangle
- zmultipoints.shp: added a large multipoint shape, followed by a
  single point.
- zlines.shp: added a line with a lot of points, followed by a short
  line.

Note the test data currently has both a polygonz.shp (single shape) and
zpolygons.shp (multiple shapes), and likewise for lines and
multipoints. This problem needed multiple shapes present in the
shapefile, so I've amended those files. I've also added them back to
the test case assertions, as the test code was no longer actually
using them.