[GEOT-6859] Speed up geotools DbaseFileWriter.FieldFormatter.getFieldString(int size, String s)
Shapefile data is stored in dbase file format and text columns have a limit of 254 bytes.

getFieldString(int size, String s) is the method responsible for truncating an
input string to one that is small enough to fit into the column.

It does this by putting the string into a string buffer and then removing a character
at a time from the end, until the string fits. But for very long strings this can
take a rather long time. The reason that characters are removed one at a time is that
the limit is in bytes, and many charsets require mutliple bytes per character.

Careful reading of the StringBuffer.replace javadocs says that the sequence (buffer)
will be 'lengthened to accommodate the specified String if necessary' hence the buffer
becoming far bigger than expected.

Solution: restructure getFieldString() so that it:
- starts with an empty buffer
- adds string to buffer (no more than size characters)
- removes excess characters until buffer byte size is no more than size
- adds any padding required

Also changed the tests so they check the strings that are returned and added more.