[GEOT-7061] GrowableInternationalString.getLocales() is fixed to the values of its first invocation
`GrowableInternationalString.getLocales():Set<Locale>`
will always return the values it had when it was first called,
regardless of having called
`GrowableInternationalString.add(Locale, String)` afterwards.

This is because the set of Locales is held in an instance variable
initialized when `getLocales()` is first called, and never updated.

Moreover, it doesn't even make sense to cache them.