[GEOT-2063] PostGIS has a longitude first axis order
PostGIS stores geometry with a XY (longitude first) axis order.

But the PostGISDialect is currently not using forceXY when
decoding the CRS. This results in a CRS with incorrect axis
ordering unless the forceXY system settings have been applied.

This commit:
- updates SQLDialect to:
  - have a forceLongitudeFirst setting
  - added constructors to set (defaults to forceLongitudeFirst off)
  - createCRS() updated to use forceLongitudeFirst setting when calling
    CRS.decode()
  - Construct PostGISDialect with forceLongitudeFirst on.

This approach was chosen to allow for better re-use if any other JDBC
implementations also require forceLongitudeFirst to be used.