[GEOT-2063] PostGIS xy axis order (#3873)
* JDBC Tests, make JDBCTestSupport aware of CRS axis order

A future change will be changing the PostGIS CRS axis order to
be longitude first to match PostGIS expectations. But there are
quite a few tests that will need to be updated.

This commit:
- makes JDBCTestSupport:
  - aware of the required axis order with a forceLongitudeFirst parameter
    (this parameter moved from JDBCDataStoreAPIOnlineTest)
  - adds method to build CRS taking account of the axis order
    - decodeEPSG(int epsgCode) to do the actual decoding
- JDBCFeatureSourceOnlineTest.getWGS84() is removed.
  Main reason for removing is that WGS84 implies a lon/lat axis order but
  this method will only provide that if the force lon/lat system settings
  are set.
- other tests are updated to use decodeEPSG() rather than decoding the
  CRS themselves. This is to allow sub class tests to control the axis
  order by setting the forceLongitudeFirst in their constructors.

Note this commit is use a refactor to make future test changes a little
easier. There is no functional changes in this commit at all.

* [GEOT-2063] PostGIS has a longitude first axis order

PostGIS stores geometry with a XY (longitude first) axis order.

But the PostGISDialect is currently not using forceXY when
decoding the CRS. This results in a CRS with incorrect axis
ordering unless the forceXY system settings have been applied.

This commit:
- updates SQLDialect to:
  - have a forceLongitudeFirst setting
  - added constructors to set (defaults to forceLongitudeFirst off)
  - createCRS() updated to use forceLongitudeFirst setting when calling
    CRS.decode()
  - Construct PostGISDialect with forceLongitudeFirst on.

This approach was chosen to allow for better re-use if any other JDBC
implementations also require forceLongitudeFirst to be used.