[GEOT-7207] Use parameters instead of hard-coded values for BBOX filters
When adding a BBOX filter to a query, the HANA plugin will hard-code the
coordinates of the bounding box instead of using parameters (?) in the
generated query.

When executing the same query repeatedly with only a different bounding
box, the database cannot reuse a plan from the SQL plan cache due to the
changed hard-coded bounding box values, but has to generate a new plan.
This can be costly, especially if the query is complex.

Hence, when adding a BBOX filter, the filter should use parameters
instead of hard-coded values.