[GEOT-7286] Let SQLDialect decide whether to apply hints on virtual tables
So far, JDBCDataStore does not add select-hints to queries on virtual
tables. This makes sense for some SQLDialects, but not for all.

This change adds a new method applyHintsOnVirtualTables() to the
SQLDialect class that allows the concrete SQLDialect to control whether
hints are added to queries on virtual tables. The default implementation
of applyHintsOnVirtualTables() returns false to keep the current
behavior.

For the HANA plugin, the behavior has been adapted to add select-hints
to queries on virtual tables.