[GEOT-7333] The result of linearization after creating circulararc is wrong
We improve the precision of the center computed from three control
points in the CircularArc class.

The improvement considers the matrix condition number and selects the
best 2 out of 3 equations for finding the center.

We reduce the error further by solving the equation system with an LU-
decomposition with pivoting.