[GEOT-7332] Resolve incompatibility between LGPL-2.1 and LGPL-3+
Before this change, the file GeometrySmoother.java was licensed both
with "LGPL-2.1 only" and "LGPL-3.0 or any later version", which is
impossible to fulfill at the same time -- which license could be used
for this file?

By changing "LGPL-2.1"to "LGPL-2.1 or any later version", it is
possible to license this file under LGPL-3.0, hereby resolving this
conflict.