[GEOT-7607] Error using function Interpolate in COLOR mode from SLD
Make `InterpolateFunction.evaluate(Object, Class)` adhere to the interface
contract by returning a value converted to the requested "context" type.

This was not true for the case when the `method` parameter is `COLOR`,
and hence the function would have failed instead of returning a
converted value or `null`, according to `Converters.covert()`.