[GEOT-7759] Interpolate properties to dependency versions in BOM files
Use flatten-maven-plugin with flattenMode=ossrh instead of bom, and
explicitly instruct it to preserve  the dependencyManagement section
instead. This results in properties being interpolated to depdendencies,
and removes the properties section. Otherwise the BOMs inherit the
properties defined in the root pom, and doesn't interpolate properties
to versions, which is not desirable.