[GEOT-7777] Add a downstream QA build
Add a separate build job to the GeoTools github actions workflows to run the QA
builds for GeoServer and GeoWebcache whenever the gt-platform-dependencies BOM is modified.

With the usage of the geotools platform-dependencies BOM in GeoServer and GeoWebcache
to manage shared dependencies, a timing issue is introduced in these projects to detect
deprecated API usage when a dependency is upgraded.