Update WKBReader base code from JTS
This WKBReader is a copy/extension of JTS's version because it's not really
practical to extend the class with additional functionality.

But the current base reader code is quite dated. In this commit we update
the base code to match the latest from JTS.

The next commit will add the geotools support for curved geometries back