[GEOT-7834] Update GeoPackage reader to support curved geometry types
* Update WKBReader base code from JTS

This WKBReader is a copy/extension of JTS's version because it's not really
practical to extend the class with additional functionality.

But the current base reader code is quite dated. In this commit we update
the base code to match the latest from JTS.

The next commit will add the geotools support for curved geometries back

* WKBReader, add the curved geometry support back in

* WKBReader code format

Done as a separate commit so the changes in the two earlier commits
are clearer.

* Add WKBReaderTest

- most example bytes were generated using PostGIS to ensure
  this reader is still good for PostGIS
- Manually tweaked the geometry code in canReadPointWithZM_OGC to make
  it fit the (ISO/OGC 06-103r4) spec

* GeometryTypeConverterFactory, fix conversion for Polygon to CurvePolygon

Currently you get a CompoundCurve when attempting to convert
Polygon to CurvePolygon

* GeoPackage, add curved geometry support to reader

* Parameterize the WKBReaderTest for PostGIS and OGC style geometry types

Parameterizing the test let's both formats be tested with less effort.
Only downside is that SRID is only supported for PostGIS but that is easy enough
to deal with.